(connection:store) enforce JSON compatible values
This removes raw strings and as a side effect, operations like append
and prepend.
What this prevents is writing a raw string that can be interpreted as
JSON when the database key is read.
Won’t effect existing raw strings, they’ll be upgraded next time the
key is modified.